Move load-all stream orchestration out of transport layer.
Wire serialization stays in transport; stream_load_all_turns lives in
api.services.load_all_streaming for Core and BFF routers to import.
