+++ packages/api/api/routers/games.py
@@ -14,6 +14,7 @@ from api.services.deps import (
     get_turn_load_service,
 )
 from api.services.game_service import GameService
+from api.services.load_all_streaming import stream_load_all_turns
 from api.services.load_all_turns import LoadAllTurnsService
 from api.services.turn_analytic_service import TurnAnalyticService
 from api.services.turn_load_service import TurnLoadService
@@ -29,11 +30,7 @@ from api.transport.connections_options import (
     FlareConnectionMode,
 )
 from api.transport.game_info_update import GameInfoUpdateRequest, RefreshGameInfoParams
-from api.transport.load_all_turns import (
-    LoadAllTurnsRequest,
-    LoadAllTurnsStatusResponse,
-    stream_load_all_turns,
-)
+from api.transport.load_all_turns import LoadAllTurnsRequest, LoadAllTurnsStatusResponse
 
 router = APIRouter(prefix="/v1/games", tags=["games"])
